refactor: Organize components into components dir
