--- lib/src/common/loading_view.dart
@@ -1,16 +1,0 @@
-import 'package:flutter/material.dart';
-import 'package:mdigits/src/common/default_appbar.dart';
-
-class LoadingView extends StatelessWidget {
-  const LoadingView({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: createAppBar(context: context),
-      body: const Center(
-        child: CircularProgressIndicator(),
-      ),
-    );
-  }
-}
