+++ lib/src/components/common/default_text.dart
