+++ lib/src/components/models/trial_data.dart
