+++ lib/src/components/response/response_widget.dart
@@ -1,0 +1,55 @@
+import 'package:flutter/material.dart';
+import 'package:flutter/services.dart';
+import 'package:mdigits/src/components/response/response_controller.dart';
+import 'package:mdigits/src/components/common/centeredbox.dart';
+import 'package:mdigits/src/components/common/default_text.dart';
+import 'package:mdigits/src/components/common/spacing_holder.dart';
+import 'package:mdigits/src/components/common/styles.dart';
+
+class ResponseWidget extends StatelessWidget {
+  const ResponseWidget({
+    Key? key,
+    required ResponseController responseController,
+  })  : _responseController = responseController,
+        super(key: key);
+
+  final ResponseController _responseController;
+
+  @override
+  Widget build(BuildContext context) {
+    return CenteredBox(
+      column: Column(
+        mainAxisAlignment: MainAxisAlignment.center,
+        children: <Widget>[
+          const DefaultText(
+            '¿Números?',
+          ),
+          const BetweenWidgetsSpace(),
+          TextField(
+            controller: _responseController.textController,
+            keyboardType: TextInputType.number,
+            inputFormatters: [FilteringTextInputFormatter.digitsOnly],
+            decoration: textFieldStyle,
+            onSubmitted: (_) => submitResponse(),
+            autofocus: true,
+          ),
+          const BetweenWidgetsSpace(),
+          ElevatedButton(
+            onPressed: () {
+              submitResponse();
+            },
+            child: Text(
+              'Seguir',
+              style: Theme.of(context).textTheme.button,
+            ),
+          ),
+        ],
+      ),
+    );
+  }
+
+  void submitResponse() {
+    _responseController.submit();
+    _responseController.toNextScreen();
+  }
+}
