--- lib/src/response/response_view.dart
@@ -1,23 +1,0 @@
-import 'package:flutter/material.dart';
-import 'package:get/get.dart';
-import 'package:mdigits/src/response/response_widget.dart';
-import 'package:mdigits/src/response/response_controller.dart';
-
-/// Screen for collecting response from participant
-class ResponseView extends StatelessWidget {
-  final ResponseController _responseController = Get.put(ResponseController());
-
-  ResponseView({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return WillPopScope(
-      onWillPop: () async {
-        return false;
-      },
-      child: Scaffold(
-        body: ResponseWidget(responseController: _responseController),
-      ),
-    );
-  }
-}
